+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - Fixed issue where lengthy lines of text could exceed the width of the build description box.
+- Fixed Bottom Heavy mutator description.
+- Fixed Prophecy mutator description.
 
 ## 2024-05-01